feat: generate the artifacts
