+++ artifacts/src/abis/index.ts
@@ -1,5 +1,6 @@
-export * from './admin-plugin-abis';
-export * from './multisig-plugin-abis';
-export * from './token-voting-plugin-abis';
-export * from './staged-proposal-processor-plugin-abis';
-export * from './osx-abis';
+
+export * from './osx/abis';
+export * from './admin-plugin/abis';
+export * from './multisig-plugin/abis';
+export * from './token-voting-plugin/abis';
+export * from './staged-proposal-processor-plugin/abis';
